Added support for serverid
A serverid var which was always set to 0. This would result in stats not showing up for server with no server with an ID of 0.

TODO: Add support to combine stats if the serverid is not set in the query. I would have implemented it but this is just a quick fix.